+++ src/main/java/com/loja_airsoft/app/repositories/ClienteRepository.java
@@ -10,5 +10,5 @@ import com.loja_airsoft.app.entities.Cliente;
 @Repository
 public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
 
-	Cliente findByCpfCliente(Integer cli_cpf);
+	Cliente findByCpfCliente(Double cli_cpf);
 }
